Add basic file structure.
Make a clean filetree with minimalist tools.
Add React and ReactDOM and add Webpack for compilation and
Webpack-dev-server for development environment in order to get benefits
of hot module replacement.
Add Sass support for stylesheets.
Add Eslint with minimal configuration and Editorconfig in order to keep
an eye of coding
style.
Add .nvmrc to specify which version of node is currently used in the
project.